+++ cli/examples/custom-command/main.rs
@@ -14,7 +14,8 @@
 
 use std::io::Write as _;
 
-use jj_cli::cli_util::{CliRunner, CommandError, CommandHelper};
+use jj_cli::cli_util::{CliRunner, CommandHelper};
+use jj_cli::command_error::CommandError;
 use jj_cli::ui::Ui;
 
 #[derive(clap::Parser, Clone, Debug)]
