+++ cli/src/commands/files.rs
@@ -16,7 +16,8 @@ use std::io::Write;
 
 use tracing::instrument;
 
-use crate::cli_util::{CommandError, CommandHelper, RevisionArg};
+use crate::cli_util::{CommandHelper, RevisionArg};
+use crate::command_error::CommandError;
 use crate::ui::Ui;
 
 /// List files in a revision
