+++ cli/src/commands/init.rs
@@ -20,7 +20,8 @@ use jj_lib::workspace::Workspace;
 use tracing::instrument;
 
 use super::git;
-use crate::cli_util::{user_error_with_hint, user_error_with_message, CommandError, CommandHelper};
+use crate::cli_util::CommandHelper;
+use crate::command_error::{user_error_with_hint, user_error_with_message, CommandError};
 use crate::ui::Ui;
 
 /// Create a new repo in the given directory
