+++ cli/src/commands/unsquash.rs
@@ -17,7 +17,8 @@ use jj_lib::object_id::ObjectId;
 use jj_lib::rewrite::merge_commit_trees;
 use tracing::instrument;
 
-use crate::cli_util::{user_error, CommandError, CommandHelper, RevisionArg};
+use crate::cli_util::{CommandHelper, RevisionArg};
+use crate::command_error::{user_error, CommandError};
 use crate::description_util::combine_messages;
 use crate::ui::Ui;
