+++ cli/src/commands/version.rs
@@ -16,7 +16,8 @@ use std::io::Write;
 
 use tracing::instrument;
 
-use crate::cli_util::{CommandError, CommandHelper};
+use crate::cli_util::CommandHelper;
+use crate::command_error::CommandError;
 use crate::ui::Ui;
 
 /// Display version information
