+++ cli/src/lib.rs
@@ -16,6 +16,7 @@
 
 pub mod cleanup_guard;
 pub mod cli_util;
+pub mod command_error;
 pub mod commands;
 pub mod commit_templater;
 pub mod config;
